CB-14578 Implement flow steps for FreeIPA CCM Upgrade
Handle CCM1 and CCM2 originating tunnel types, the latter is just some simple update with
the rest of the steps are skipped.

CCMv1 migration covers setting registering a new inverting proxy host and agent,
pushing the metadata via Salt pillars and applying the upgrade step.

The upgrade ccm log file is also collected with the log bundle.

The upgrade itself runs network connectivity check, the re-registers Cluster Proxy,
finally disables and deregisters Mina client (CCMv1).

This commit will close the following tickets too:
CB-14579 CB-14580 CB-14581 CB-14583 CB-14586 CB-14588 CB-14592 CB-15870

N.B. This change does not deal with rolling back failed upgrade, that will be
covered in an upcoming commit.

+++ freeipa/src/main/java/com/sequenceiq/freeipa/flow/stack/upgrade/ccm/handler/UpgradeCcmRemoveMinaHandler.java
@@ -11,6 +11,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import com.sequenceiq.cloudbreak.common.event.Selectable;
+import com.sequenceiq.cloudbreak.orchestrator.exception.CloudbreakOrchestratorException;
 import com.sequenceiq.flow.reactor.api.handler.HandlerEvent;
 import com.sequenceiq.freeipa.flow.stack.upgrade.ccm.UpgradeCcmService;
 import com.sequenceiq.freeipa.flow.stack.upgrade.ccm.event.UpgradeCcmEvent;
@@ -33,15 +34,24 @@ public class UpgradeCcmRemoveMinaHandler extends AbstractUpgradeCcmEventHandler 
 
     @Override
     protected Selectable defaultFailureEvent(Long resourceId, Exception e, Event<UpgradeCcmEvent> event) {
-        LOGGER.error("Removing mina for CCM upgrade has failed", e);
+        LOGGER.error("Removing Mina for CCM upgrade has failed", e);
         return new UpgradeCcmFailureEvent(UPGRADE_CCM_FAILED_EVENT.event(), resourceId, e);
     }
 
     @Override
     protected Selectable doAccept(HandlerEvent<UpgradeCcmEvent> event) {
         UpgradeCcmEvent request = event.getData();
-        LOGGER.info("Remove mina for CCM upgrade...");
-        upgradeCcmService.removeMina(request.getResourceId());
+        if (request.getOldTunnel().useCcmV1()) {
+            LOGGER.info("Remove Mina for CCM upgrade...");
+            try {
+                upgradeCcmService.removeMina(request.getResourceId());
+            } catch (CloudbreakOrchestratorException e) {
+                LOGGER.debug("Failed removing Mina service with a salt state");
+                return new UpgradeCcmFailureEvent(UPGRADE_CCM_FAILED_EVENT.event(), request.getResourceId(), e);
+            }
+        } else {
+            LOGGER.info("Remove Mina step is skipped for previous tunnel type '{}'", request.getOldTunnel());
+        }
         return UPGRADE_CCM_REMOVE_MINA_FINISHED_EVENT.createBasedOn(request);
     }
